WebP

Description

WebP decoding and encoding functions.

The inputs of encoding functions and outputs of decoding functions are arrays of bytes representing pixels, with one byte per channel.

For example functions operating on arrays in RGB layout expect them to be in the following format:

import qualified Data.ByteString as Bytes
import Data.Word

Bytes.pack ([r, g, b, r, g, b, ..] :: Word8)

Documentation

module WebP.Types

decodeRGB :: ByteString -> ByteString

Deocde an array of bytes in RGB layout.

decodeBGR :: ByteString -> ByteString

Deocde an array of bytes in BGR layout.

decodeRGBA :: ByteString -> ByteString

Deocde an array of bytes in RGBA layout.

decodeBGRA :: ByteString -> ByteString

Deocde an array of bytes in BGRA layout.

getInfo :: ByteString -> Maybe (Width, Height)

Validate the WebP image header and retrieve the image width and height.

newtype Width

Represents width of an image.

Constructors

Width Int

newtype Height

Represents height of an image.

Constructors

Height Int

newtype Quality

Represents quality to be used in encoding operations. Valid ranges are between 0 and 100.

Constructors

Quality Float

encodeRGB :: ByteString -> Width -> Height -> Quality -> ByteString

Encode an array of bytes in RGB layout at the specified width, height and quality. Quality ranges from 0 to 100. The value 0 corresponds to low quality and small output sizes, whereas 100 is the highest quality and largest output size.

encodeBGR :: ByteString -> Width -> Height -> Quality -> ByteString

Encode an array of bytes in BGR layout at the specified width, height and quality. Quality ranges from 0 to 100. The value 0 corresponds to low quality and small output sizes, whereas 100 is the highest quality and largest output size.

encodeRGBA :: ByteString -> Width -> Height -> Quality -> ByteString

Encode an array of bytes in RGBA layout at the specified width, height and quality. Quality ranges from 0 to 100. The value 0 corresponds to low quality and small output sizes, whereas 100 is the highest quality and largest output size.

encodeBGRA :: ByteString -> Width -> Height -> Quality -> ByteString

Encode an array of bytes in BGRA layout at the specified width, height and quality. Quality ranges from 0 to 100. The value 0 corresponds to low quality and small output sizes, whereas 100 is the highest quality and largest output size.

losslessEncodeRGB :: ByteString -> Width -> Height -> ByteString

Lossless variant of encodeRGB

losslessEncodeBGR :: ByteString -> Width -> Height -> ByteString

Lossless variant of encodeBGR

losslessEncodeRGBA :: ByteString -> Width -> Height -> ByteString

Lossless variant of encodeRGBA

losslessEncodeBGRA :: ByteString -> Width -> Height -> ByteString

Lossless variant of encodeBGRA